[CHORE] VC 네이밍 변경 및 View 관련 코드 뷰 내부로 이동(#434)
