+++ Manito/Manito/Screens/Interaction/View/SelectManitteeView.swift
@@ -115,6 +115,45 @@ final class SelectManitteeView: UIView {
         self.joystickBackgroundView.addGestureRecognizer(swipeRight)
     }
 
+    private func setupGifImage() {
+        switch stageType {
+        case .showJoystick:
+            DispatchQueue.main.async {
+                self.joystickImageView.animate(withGIFNamed: ImageLiterals.gifJoystick)
+            }
+        case .showCapsule:
+            self.joystickImageView.stopAnimatingGIF()
+            DispatchQueue.main.async {
+                self.openCapsuleImageView.animate(withGIFNamed: ImageLiterals.gifCapsule, loopCount: 1, animationBlock: { [weak self] in
+                    self?.stageType = .openName
+                })
+            }
+        case .openName:
+            self.openCapsuleImageView.stopAnimatingGIF()
+            DispatchQueue.main.asyncAfter(deadline: .now() + 0.5, execute: {
+                self.stageType = .openButton
+            })
+        case .openButton:
+            break
+        }
+    }
+
+    private func hiddenImageView() {
+        switch stageType {
+        case .showJoystick:
+            self.nameLabel.alpha = 0.0
+            self.openCapsuleImageView.isHidden = true
+            self.confirmButton.isHidden = true
+        case .showCapsule:
+            self.openCapsuleImageView.isHidden = false
+            self.joystickBackgroundView.isHidden = true
+        case .openName:
+            self.nameLabel.fadeIn()
+        case .openButton:
+            self.confirmButton.isHidden = false
+        }
+    }
+
     func configureUI(manitteeNickname: String) {
         self.nameLabel.text = manitteeNickname
     }
